moved agi value to the agilityvalue module
some issues remain
